ci: merge axiom validation checks
