+++ AXIOMS.md
@@ -100,7 +100,7 @@ Wrapping modular arithmetic at 2^256 is **proven**, not assumed. All 15 pure bui
 
 - Active axioms: 1
 - Production blockers from axioms: 0
-- Enforcement: `scripts/check_axiom_locations.py` ensures this file tracks exact source location.
+- Enforcement: `scripts/check_axioms.py` ensures this file tracks exact source location.
 - Compilation-path totalization work in `Compiler/CompilationModel.lean` does not
   add, remove, or modify Lean axioms; it only replaces `partial def` recursion
   with explicit structural termination proofs (including dynamic-param scope
