+++ Makefile
@@ -76,7 +76,7 @@ profile-lean: ## Profile Lean module build time and update docs/LEAN_PERF_QUEUE.
 
 axiom-report: ## Generate axiom dependency report for all 550 theorems
 	lake env lean PrintAxioms.lean 2>&1 | tee axiom-report-raw.log
-	python3 scripts/check_axiom_report.py --log axiom-report-raw.log
+	python3 scripts/check_axioms.py --log axiom-report-raw.log
 
 # ---------------------------------------------------------------------------
 # Compilation
@@ -106,7 +106,7 @@ check: ## Run local CI-equivalent checks job (no Lean build, no solc)
 	python3 scripts/check_property_coverage.py
 	python3 scripts/check_contract_structure.py
 	python3 scripts/check_paths.py
-	python3 scripts/check_axiom_locations.py
+	python3 scripts/check_axioms.py
 	python3 scripts/generate_verification_status.py --check
 	python3 scripts/check_verification_status_doc.py
 	python3 scripts/check_interop_matrix_sync.py
