+++ docs-site/content/add-contract.mdx
@@ -111,7 +111,7 @@ python3 scripts/check_storage_layout.py
 python3 scripts/check_selectors.py
 python3 scripts/check_contract_structure.py
 python3 scripts/generate_verification_status.py --check
-python3 scripts/check_axiom_locations.py
+python3 scripts/check_axioms.py
 python3 scripts/check_lean_hygiene.py
 ```
